api/requirements.txt
psycopg2-binary
gunicorn
djangorestframework
Django
uvicorn
django-elasticsearch-dsl
django_elasticsearch_dsl_drf